Italian localization (extended) [Requires Feedback]
This is my first version of an alternative localization into Italian language which is a follow-up of the current localization file made by @EugenioBuffo (#3725), which I thanks, and of my discussion "Italian localization (git newbie)" (#3633) which covers the main user interface, all the current the Extensions and Scripts, with the following exceptions:

    txt2img2img (I got errors therefore I removed it from my local installation of SD Web UI) 
    Parameter Sequencer (not installed locally)
    Booru tag autocompletion (not installed locally)
    Saving steps of the sampling process (not installed locally)
	
I do not forecast to translate the above scripts in the short period, unless I will install them locally on my machine.

I beg your pardon if I am brutally overwriting the originally submitted file but I find quite exhausting to edit and append over a thousand lines of code to the original file. If this is mandatory, then I will delete this commit and start a new one amending the original it_IT.json file.

It is for sure not perfect and there are some translations that can be improved, therefore I wish to invite @EugenioBuffo and any other Italian mother language person willing give advice and to help to review this extensive translation . I look forward read any feedback from the community and developers. Thank you.